refactor(client): change style to camelCase
