+++ packages/client/src/assets/fonts/rubicBubblesCyrillic.woff2
